ai_assistant_cdea/requirements.txt
flask==3.1.0
flask-cors==6.0.1
flask-socketio==5.5.1
python-dotenv==1.1.1
anthropic==0.57.1
pandas==2.3.1
openpyxl==3.1.5
python-socketio==5.13.0
python-engineio==4.12.2
eventlet==0.33.3
httpx==0.28.1
jinja2==3.1.6
werkzeug==3.1.3